Adiciona os parametros de autenticacao na query string

+++ home/src/main/java/com/mobsky/home/data/network/api/model/comics/Price.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Price(
     @SerializedName("price")
-    val price: Int? = 0,
+    val price: Double? = 0.0,
     @SerializedName("type")
     val type: String? = ""
 )
